services/nlp/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
httpx>=0.28.0
textrazor>=1.4.1
python-dotenv>=1.0.0
